Refactor settings and utilities: update order settings, add list and string utility functions
