+++ src/enum/payment-status.ts
@@ -1,9 +1,9 @@
 export enum PaymentStatus {
-  NOT_PAID = 'not_paid',
-  AWAITING = 'awaiting',
-  CAPTURED = 'captured',
-  PARTIALLY_REFUNDED = 'partially_refunded',
-  REFUNDED = 'refunded',
-  CANCELED = 'canceled',
-  REQUIRES_ACTION = 'requires_action',
+	NOT_PAID = 'not_paid',
+	AWAITING = 'awaiting',
+	SUCCESS = 'success',
+	PARTIALLY_REFUNDED = 'partially_refunded',
+	REFUNDED = 'refunded',
+	CANCELED = 'canceled',
+	REQUIRES_ACTION = 'requires_action',
 }
