+++ src/utils/index.ts
@@ -1,1 +1,2 @@
-export * from './list_utils';
+export * from './list-utils';
+export * from './string-utils';
